moved variables into their own file, from core.go
